Updated [BAEL-7632] Convert a phone number in words to number with Java
